requirements.txt
# Wildlife Detector AI v2.0 Dependencies
# Python 3.12.x required for SpeciesNet 5.0 compatibility

# Core AI/ML dependencies
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
opencv-python==4.9.0.80

# SpeciesNet (to be installed separately due to specific requirements)
# Installation: pip install git+https://github.com/microsoft/SpeciesNet.git

# GUI Framework
PySide6==6.7.1
pyqtgraph==0.13.7

# Utilities
pyyaml==6.0.1
click==8.1.7
structlog==24.2.0
python-dotenv==1.0.1

# Image processing
scikit-image==0.23.2
imageio==2.34.1

# Data export
openpyxl==3.1.2
xlsxwriter==3.2.0
matplotlib==3.9.0
seaborn==0.13.2

# Development tools
pytest==8.2.2
pytest-cov==5.0.0
pytest-qt==4.4.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1

# Type stubs
types-pyyaml==6.0.12.20240311
types-requests==2.32.0.20240602

# Packaging
setuptools==70.0.0
wheel==0.43.0
pyinstaller==6.7.0

# Performance monitoring
psutil==5.9.8
memory-profiler==0.61.0

# Async support
asyncio==3.4.3
aiofiles==23.2.1
